Create userspace ebpf tracer code
This code is not complete and still presents some bugs, but it
implements most of the expected logic from userspace.

Most of the work will be done in the kernel programs which are yet to be
written, but this commit ensures that the programs can be loaded
successfully (the attack call is still commented).

There are still things to figure out:
  - How to stop the Mevi progam when the tracer is done.
  - Figure out the corresponce between ptrace-uffd events and ebpf
    programs.
  - Work on move some abstraction to the `mevi-driver` crate.